optimize-service/requirements.txt
# Optimize Service - Python Dependencies

# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Optimization
pyomo==6.7.0
highspy==1.5.3  # HiGHS solver (open-source, fast)

# Scientific computing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# Task queue (async optimization)
redis==5.0.1
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-prometheus==0.43b0
prometheus-client==0.19.0

# Logging
structlog==24.1.0

# HTTP client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting and formatting
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Type stubs
types-python-dateutil==2.8.19.14
pandas-stubs==2.1.4.231227
types-redis==4.6.0.20240106
